Finish Project: Landing Page
It was quite a ride. Learned a lot. Struggled with images inside div flex items, turns out, need to set flex to "flex: 0 1 0;" to generate positive space for justify-content to work. Afterwards, I set a dedicated max width/height to the images and it worked. Used media query as well to make website more responsive.